put all resource tests together and made the RUD only RU to keep it self contained
