+++ src/test/elements/zendesk/assets/categories.json
@@ -2,5 +2,6 @@
   "name": "Sample Category 1",
   "description": "A Category Created via API",
   "position": 2,
-  "locale": "us-en"
+  "locale": "us-en",
+  "body": "<<random.uuid>>"
 }
